+++ go/azcosmoscx/include/cosmoscx.h
@@ -374,7 +374,7 @@ struct CosmosCxFfiResult_Pipeline cosmoscx_v0_query_pipeline_create(CosmosCxStr 
 struct CosmosCxFfiResult_Pipeline cosmoscx_v0_readmany_pipeline_create(CosmosCxStr item_identities,
                                                                        CosmosCxStr pkranges,
                                                                        CosmosCxStr pk_kind,
-                                                                       uint32_t pk_version);
+                                                                       uint8_t pk_version);
 
 /**
  * Frees the memory associated with a pipeline.
